Fix some compiler warnings (implicit casts etc) in ugens
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@7789 a380766d-ff14-0410-b294-a243070f3f08

+++ Source/plugins/GrainUGens.cpp
@@ -123,7 +123,7 @@ struct WarpWinGrain
 
 struct Warp1 : public Unit
 {
-	int m_fbufnum;
+	float m_fbufnum;
 	SndBuf* m_buf;
 	int mNumActive[16];
 	int mNextGrain[16];
